Moves charts to config to delete chart.js
